etf_web/requirements.txt
Flask==2.3.2
Flask-Limiter==3.5.0
python-dotenv==1.0.0